+++ src/components/index.ts
@@ -27,6 +27,8 @@ import {
   TransfersEmptyState,
   TransferSummary,
   UserFlagsCard,
+  UserFlagsContent,
+  UserFlagsEmptyState,
   UserProfileSummary,
   UserRegistrationProgress,
   UserSearchResultsEmptyState,
@@ -77,6 +79,8 @@ export {
   TransferSummary,
   TwoColumnContentWrapper,
   UserFlagsCard,
+  UserFlagsContent,
+  UserFlagsEmptyState,
   UserProfileSummary,
   UserRegistrationProgress,
   UserSearchResultsEmptyState,
