+++ src/components/molecules/UserFlagsCard/UserFlagsCard.tsx
@@ -1,7 +1,11 @@
+import { useState } from 'react';
 import { User } from 'types';
+import UserFlagsContent from '../UserFlagsContent';
+import UserFlagsEmptyState from '../UserFlagsEmptyState';
 
 const UserFlagsCard = ({ user }: { user: User }) => {
   const hasUserFlags = user.userFlags && user.userFlags.length > 0;
+  const [collapsed, setCollapsed] = useState(false);
 
   return (
     <div className="mx-auto container bg-white shadow rounded overflow-x-scroll">
@@ -11,22 +15,33 @@ const UserFlagsCard = ({ user }: { user: User }) => {
         <h3 className="text-lg leading-6 font-medium text-gray-900">
           User Flags
         </h3>
+        {/* Collapse Button */}
+        {hasUserFlags && user!.userFlags!.length > 1 && (
+          <button
+            className="bg-blue-400 p-2 rounded-lg focus:outline-none focus:text-gray-500 transition duration-150 ease-in-out text-white text-sm"
+            onClick={() => setCollapsed(!collapsed)}
+          >
+            Collapse
+          </button>
+        )}
       </div>
       {/* Content  */}
       <div className="px-4">
         {hasUserFlags ? (
           <div>
-            {user!.userFlags!.map((userFlag, idx) => (
-              <UserFlagsContent
-                key={`userFlag-${userFlag}-${idx}`}
-                userFlag={userFlag}
-                idx={idx}
-                totalFlags={user.userFlags!.length}
-              />
-            ))}
+            {user!
+              .userFlags!.slice(0, collapsed ? 1 : user!.userFlags!.length)
+              .map((userFlag, idx) => (
+                <UserFlagsContent
+                  key={`userFlag-${userFlag}-${idx}`}
+                  userFlag={userFlag}
+                  idx={idx}
+                  totalFlags={user.userFlags!.length}
+                />
+              ))}
           </div>
         ) : (
-          <NoFlagsPlaceholder user={user} />
+          <UserFlagsEmptyState user={user} />
         )}
       </div>
     </div>
@@ -34,24 +49,3 @@ const UserFlagsCard = ({ user }: { user: User }) => {
 };
 
 export default UserFlagsCard;
-
-const UserFlagsContent = ({
-  userFlag,
-  idx,
-  totalFlags,
-}: {
-  userFlag: string;
-  idx: number;
-  totalFlags: number;
-}) => (
-  <div className={`${idx === 0 && 'mt-2'} ${idx === totalFlags - 1 && 'pb-2'}`}>
-    {idx != 0 && <hr className="my-2" />}
-    <p>{userFlag}</p>
-  </div>
-);
-
-const NoFlagsPlaceholder = ({ user }: { user: User }) => (
-  <div className="p4">
-    <p>No Flags for User: {user.userID}</p>
-  </div>
-);
