+++ src/components/molecules/UserFlagsContent/index.ts
@@ -1,0 +1,3 @@
+import UserFlagsContent from './UserFlagsContent';
+
+export default UserFlagsContent;
