+++ src/components/molecules/UserFlagsEmptyState/index.ts
@@ -1,0 +1,3 @@
+import UserFlagsEmptyState from './UserFlagsEmptyState';
+
+export default UserFlagsEmptyState;
